Rename HARKparallel to HARK.parallel
Rename the parallel module, and refer to it by its new name in all the
code in the repo that currently refers to it.

Issue #85 Add setup.py so module can be installed
