+++ HARK/ConsumptionSaving/Demos/Fagereng_demo.py
@@ -38,12 +38,13 @@ import os
 sys.path.insert(0, os.path.abspath('../'))
 sys.path.insert(0, os.path.abspath('../..'))
 sys.path.insert(0, os.path.abspath('../../cstwMPC'))
+sys.path.insert(0, os.path.abspath('../../../'))
 
 import numpy as np
 from copy import deepcopy
 
 from HARKutilities import approxUniform, getPercentiles
-from HARKparallel import multiThreadCommands
+from HARK.parallel import multiThreadCommands
 from HARKestimation import minimizeNelderMead
 from ConsIndShockModel import IndShockConsumerType
 from SetupParamsCSTW import init_infinite # dictionary with most ConsumerType parameters
